+++ src/nistapi/lvl1/api.c
@@ -1,6 +1,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 #include <api.h>
+#include <stddef.h>
 #include <sig.h>
 
 int
@@ -10,15 +11,15 @@ crypto_sign_keypair(unsigned char *pk, unsigned char *sk) {
 }
 
 int
-crypto_sign(unsigned char *sm, unsigned long long *smlen,
-            const unsigned char *m, unsigned long long mlen,
+crypto_sign(unsigned char *sm, size_t *smlen,
+            const unsigned char *m, size_t mlen,
             const unsigned char *sk) {
     return sqisign_sign(sm, smlen, m, mlen, sk);
 }
 
 int
-crypto_sign_open(unsigned char *m, unsigned long long *mlen,
-                 const unsigned char *sm, unsigned long long smlen,
+crypto_sign_open(unsigned char *m, size_t *mlen,
+                 const unsigned char *sm, size_t smlen,
                  const unsigned char *pk) {
     return sqisign_open(m, mlen, sm, smlen, pk);
 }
