+++ src/nistapi/lvl3/api.h
@@ -3,6 +3,8 @@
 #ifndef api_h
 #define api_h
 
+#include <stddef.h>
+
 #define CRYPTO_SECRETKEYBYTES 1138
 #define CRYPTO_PUBLICKEYBYTES   96
 #define CRYPTO_BYTES           263
@@ -13,13 +15,13 @@ int
 crypto_sign_keypair(unsigned char *pk, unsigned char *sk);
 
 int
-crypto_sign(unsigned char *sm, unsigned long long *smlen,
-            const unsigned char *m, unsigned long long mlen,
+crypto_sign(unsigned char *sm, size_t *smlen,
+            const unsigned char *m, size_t mlen,
             const unsigned char *sk);
 
 int
-crypto_sign_open(unsigned char *m, unsigned long long *mlen,
-                 const unsigned char *sm, unsigned long long smlen,
+crypto_sign_open(unsigned char *m, size_t *mlen,
+                 const unsigned char *sm, size_t smlen,
                  const unsigned char *pk);
 
 #endif /* api_h */
