+++ src/sqisign.c
@@ -16,8 +16,8 @@ int sqisign_keypair(unsigned char *pk, unsigned char *sk) {
 }
 
 int sqisign_sign(unsigned char *sm,
-              unsigned long long *smlen, const unsigned char *m,
-              unsigned long long mlen, const unsigned char *sk) {
+              size_t *smlen, const unsigned char *m,
+              size_t mlen, const unsigned char *sk) {
     int ret = 0;
     secret_key_t skt;
     public_key_t pkt = { 0 };
@@ -38,8 +38,8 @@ int sqisign_sign(unsigned char *sm,
 }
 
 int sqisign_open(unsigned char *m,
-              unsigned long long *mlen, const unsigned char *sm,
-              unsigned long long smlen, const unsigned char *pk) { 
+              size_t *mlen, const unsigned char *sm,
+              size_t smlen, const unsigned char *pk) { 
     int ret = 0;
     public_key_t pkt = { 0 };
     signature_t sigt;
@@ -60,8 +60,8 @@ int sqisign_open(unsigned char *m,
 }
 
 int sqisign_verify(const unsigned char *m,
-                unsigned long long mlen, const unsigned char *sig,
-                unsigned long long siglen, const unsigned char *pk) {
+                size_t mlen, const unsigned char *sig,
+                size_t siglen, const unsigned char *pk) {
 
     int ret = 0;
     public_key_t pkt = { 0 };
